rbc_telegram_collector/requirements.txt
telethon>=1.35.0
pydantic>=2.6.0
pydantic-settings>=2.2.1
PyYAML>=6.0.1
python-dateutil>=2.9.0.post0
pandas>=2.1.0
